HitraceMeter接口运行线程，如果使用了HitraceChain，将HitraceId嵌入并输出到Trace数据中

+++ interfaces/native/innerkits/src/hitrace_meter.cpp
@@ -152,7 +152,7 @@ void WriteToTraceMarker(const char* buf, const int count)
 string ToHexStr(uint64_t source)
 {
     std::stringstream ss;
-    ss << std::hex << (uint64_t)source;
+    ss << std::hex << source;
     return ss.str();
 }
